Changed serialization to be byte-equal to the Zend serializer
This introduces some obvious encoding errata, mainly arryas with sparse, numeric (or biased) indezes break and are reordered in a way that all numeric keys are counted sequentially from 0.
However, as this doesn't seem to affect us the ability to compare on a byte level is more important for now.